Create units.hh

+++ src/thunk/thunk.cc
@@ -249,9 +249,9 @@ Thunk::get_allowed_files( const std::string & thunk_hash ) const
   return allowed_files;
 }
 
-off_t Thunk::infiles_size( const bool include_executables ) const
+size_t Thunk::infiles_size( const bool include_executables ) const
 {
-  off_t total_size = 0;
+  size_t total_size = 0;
 
   for ( const InFile & infile : infiles_ ) {
     if ( infile.type() == InFile::Type::FILE or
